Improves core module packages.
