--- core/src/main/kotlin/chat/rocket/core/internal/model/SocketToken.kt
@@ -1,12 +1,0 @@
-package chat.rocket.core.internal.model
-
-import chat.rocket.common.internal.ISO8601Date
-import com.squareup.moshi.Json
-import se.ansman.kotshi.JsonSerializable
-
-@JsonSerializable
-data class SocketToken(
-    @Json(name = "id") val userId: String,
-    @Json(name = "token") val authToken: String,
-    @Json(name = "tokenExpires") @ISO8601Date val expiresAt: Long?
-)
