+++ core/src/main/kotlin/chat/rocket/core/internal/realtime/socket/Login.kt
@@ -1,8 +1,19 @@
-package chat.rocket.core.internal.realtime
+package chat.rocket.core.internal.realtime.socket
 
-import chat.rocket.core.internal.model.SocketToken
+import chat.rocket.common.model.Token
+import chat.rocket.core.internal.realtime.socket.model.SocketToken
 import chat.rocket.core.internal.model.TypedResponse
+import chat.rocket.core.internal.realtime.message.loginMethod
 import com.squareup.moshi.Types
+
+fun Socket.login(token: Token?) {
+    token?.let { authToken ->
+        socket?.let {
+            setState(State.Authenticating())
+            send(loginMethod(generateId(), authToken.authToken))
+        }
+    }
+}
 
 internal fun Socket.processLoginResult(text: String) {
     val type = Types.newParameterizedType(TypedResponse::class.java, SocketToken::class.java)
