+++ core/src/main/kotlin/chat/rocket/core/internal/realtime/socket/model/StramMessage.kt
@@ -1,0 +1,9 @@
+package chat.rocket.core.internal.realtime.socket.model
+
+data class StreamMessage<out T>(val type: Type, val data: T)
+
+sealed class Type {
+    object Inserted : Type()
+    object Updated : Type()
+    object Removed : Type()
+}
